LayoutTests/storage/indexeddb/modern/date-basic-expected.txt

This tests using Date objects as keys and values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Initial upgrade versionchange transaction complete
ALERT: Success getting key 'Fri Nov 04 1955 17:00:00 GMT-0700 (PDT)' of type object, result is 'Flux capacitor' of type string
ALERT: Key is a Date object, btw
ALERT: Success getting key 'Sat Nov 12 1955 10:00:00 GMT-0800 (PST)' of type object, result is 'Fish under the sea' of type string
ALERT: Key is a Date object, btw
ALERT: Success getting key 'Wed Oct 21 2015 09:00:00 GMT-0700 (PDT)' of type object, result is 'Hoverboards' of type string
ALERT: Key is a Date object, btw
ALERT: Success getting key 'a' of type string, result is 'Fri Nov 04 1955 17:00:00 GMT-0700 (PDT)' of type object
ALERT: Result is a Date object, btw
ALERT: Success getting key 'b' of type string, result is 'Sat Nov 12 1955 10:00:00 GMT-0800 (PST)' of type object
ALERT: Result is a Date object, btw
ALERT: Success getting key 'c' of type string, result is 'Wed Oct 21 2015 09:00:00 GMT-0700 (PDT)' of type object
ALERT: Result is a Date object, btw
ALERT: readonly transaction complete
PASS successfullyParsed is true

TEST COMPLETE
